Data.Functor.Contravariant.Adjunction

Description

Documentation

class (Contravariant f, Representable g) => Adjunction f g | f -> g, g -> f where

An adjunction from Hask^op to Hask

Op (f a) b ~ Hask a (g b)

rightAdjunct unit = id
leftAdjunct counit = id

Any adjunction from Hask to Hask^op would indirectly permit unsafePerformIO, and therefore does not exist.

Minimal complete definition

unit, counit | leftAdjunct, rightAdjunct

Methods

unit :: a -> g (f a)

counit :: a -> f (g a)

leftAdjunct :: (b -> f a) -> a -> g b

rightAdjunct :: (a -> g b) -> b -> f a

adjuncted :: (Adjunction f g, Profunctor p, Functor h) => p (a -> g b) (h (c -> g d)) -> p (b -> f a) (h (d -> f c))

leftAdjunct and rightAdjunct form two halves of an isomorphism.

This can be used with the combinators from the lens package.

adjuncted :: Adjunction f g => Iso' (b -> f a) (a -> g b)

contrarepAdjunction :: Adjunction f g => (a -> f ()) -> g a

Represent a Contravariant functor that has a left adjoint

coindexAdjunction :: Adjunction f g => g a -> a -> f ()