Control.Monad.Trans.Contravariant.Adjoint

Description

Uses a contravariant adjunction:

f -| g : Hask^op -> Hask

to build a Comonad to Monad transformer. Sadly, the dual construction, which builds a Comonad out of a Monad, is uninhabited, because any Adjunction of the form

f -| g : Hask -> Hask^op

would trivially admit unsafePerformIO.

Documentation

type Adjoint f g = AdjointT f g Identity

runAdjoint :: Contravariant g => Adjoint f g a -> g (f a)

adjoint :: Contravariant g => g (f a) -> Adjoint f g a

newtype AdjointT f g w a

Constructors

AdjointT

Fields

runAdjointT :: g (w (f a))