Data.Functor.Contravariant.Rep

Description

Representable contravariant endofunctors over the category of Haskell types are isomorphic to (_ -> r) and resemble mappings to a fixed range.

Representable Contravariant Functors

class Contravariant f => Representable f where

A Contravariant functor f is Representable if tabulate and index witness an isomorphism to (_ -> Rep f).

tabulate . index ≡ id
index . tabulate ≡ id

Minimal complete definition

tabulate, index

Associated Types

type Rep f :: *

Methods

tabulate :: (a -> Rep f) -> f a

contramap f (tabulate g) = tabulate (g . f)

index :: f a -> a -> Rep f

contramapWithRep :: (b -> Either a (Rep f)) -> f a -> f b

contramapWithRep f p ≡ tabulate $ either (index p) id . f

tabulated :: (Representable f, Representable g, Profunctor p, Functor h) => p (f a) (h (g b)) -> p (a -> Rep f) (h (b -> Rep g))

tabulate and index form two halves of an isomorphism.

This can be used with the combinators from the lens package.

tabulated :: Representable f => Iso' (a -> Rep f) (f a)

Default definitions

contramapRep :: Representable f => (a -> b) -> f b -> f a